DetectionGUI/requirements.txt
Pillow==10.3.0
torch==2.2.1+cu118
torchvision==0.17.1+cu118
